Correcion de errores (#89)

+++ CarUPO/app/Http/Controllers/AccesoriosController.php
@@ -61,7 +61,9 @@ class AccesoriosController extends Controller
         $accesorio->nombre = $request->nombre;
         $accesorio->fk_producto_id = $producto->id;
         $accesorio->save();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $productos = Producto::paginate(8);
+        $success = "Accesorio creado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
 
 
@@ -144,13 +146,30 @@ class AccesoriosController extends Controller
 
         $accesorio->nombre = $request->nombre;
         $accesorio->save();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $productos = Producto::paginate(8);
+        $success = "Accesorio editado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
 
     public function eliminarAccesorio(Request $request)
     {
+
         $accesorio = Accesorio::findOrFail($request->id);
+        $producto = $accesorio->producto;
+        if ($producto->lineas_de_compra->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el accesorio porque está asociado a lineas de compra.');
+        }
+        if ($producto->lineas_de_carrito->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el accesorio porque está asociado a lineas de carrito.');
+        }
+        if ($producto->favoritos_productos->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el accesorio porque está asociada a favoritos.');
+        }
+
         $accesorio->delete();
-        return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
+        $producto->delete();
+        $productos = Producto::paginate(8);
+        $success = "Accesorio eliminado correctamente.";
+        return view('productos', compact('productos', "success"));
     }
 }
